Move PortalGeneric to lib
