+++ src/components/Popup/Popup.tsx
@@ -3,9 +3,9 @@ import * as PropTypes from 'prop-types'
 import * as _ from 'lodash'
 import CSSProperties = React.CSSProperties
 
-import { childrenExist, customPropTypes, isBrowser } from '../../lib'
+import { childrenExist, customPropTypes, isBrowser, IRenderResultConfig } from '../../lib'
 import { ItemShorthand, Extendable } from '../../../types/utils'
-import { IPortalGenericProps, IPortalGenericState, PortalGeneric } from '../Portal'
+import { IPortalGenericProps, IPortalGenericState, PortalGeneric } from '../../lib/PortalGeneric'
 import PopupContent from './PopupContent'
 
 type PopupPosition =
@@ -106,7 +106,7 @@ export default class Popup extends PortalGeneric<Extendable<IPopupState>, IPopup
 
   public state = { style: {}, open: false }
 
-  public render() {
+  public renderComponent({  }: IRenderResultConfig<IPopupProps>): React.ReactNode {
     const { basic, children, content, trigger } = this.props
     const { style } = this.state
