+++ src/components/Portal/index.ts
@@ -1,4 +1,1 @@
 export { default } from './Portal'
-export { IPortalGenericProps } from './PortalGeneric'
-export { IPortalGenericState } from './PortalGeneric'
-export { PortalGeneric } from './PortalGeneric'
